requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
firebase-admin==6.2.0
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2
python-json-logger==2.0.7
python-telegram-bot==20.7
asyncio==3.4.3
websocket-client==1.9.0
